Added Controllers

+++ OrdersApi/Models/Product.cs
@@ -4,10 +4,10 @@
     {
         public int Id { get; set; }
         public string Name { get; set; } = string.Empty;
-        public decimal Price { get; set; }
+        public double Price { get; set; }
 
 
-        public decimal CategoryId { get; set; } // Category FK
+        public int CategoryId { get; set; } // Category FK
         public Category Category { get; set; } // Product can be included in one category
         public ICollection<OrderDetail> OrderDetails { get; set; } // Product can be included more than one OrderDetails
     }
